more floats, fixed grain size

+++ src/benchmarks/benchmarks.h
@@ -32,7 +32,7 @@ template <typename DataType> inline auto distribution()
   }
 }
 
-template <typename DataType = double, typename Index = size_t>
+template <typename DataType = float, typename Index = size_t>
 std::pair<Vec<DataType>, Vec<Index> > makeData(size_t n)
 {
   Vec<DataType> vec(n);
@@ -49,7 +49,7 @@ std::pair<Vec<DataType>, Vec<Index> > makeData(size_t n)
   return make_pair(vec, index);
 }
 
-template <typename DataType = double, typename Index = size_t>
+template <typename DataType = float, typename Index = size_t>
 std::pair<Vec<DataType>, Vec<Index> >
 makeDataLocal(size_t n, size_t chunk_size = 1024, double prob = 1.0)
 {
